% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_patterns_equal}
\alias{are_patterns_equal}
\title{Check if dimension patterns are equivalent}
\usage{
are_patterns_equal(pattern1, pattern2)
}
\arguments{
\item{pattern1}{First dimension pattern}

\item{pattern2}{Second dimension pattern}
}
\value{
Logical indicating if patterns are equivalent
}
\description{
Check if dimension patterns are equivalent
}
