% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sl4x.R
\name{get_dims_strings}
\alias{get_dims_strings}
\title{Extract Unique Dimension Strings}
\usage{
get_dims_strings(sl4_obj)
}
\arguments{
\item{sl4_obj}{A structured SL4 object containing dimension information.}
}
\value{
A data frame containing unique dimension strings.
}
\description{
Extracts and lists unique dimension strings present in an SL4 object.
}
