% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dim_names}
\alias{process_dim_names}
\title{Process Dimension Names to Handle Duplicates}
\usage{
process_dim_names(dim_names)
}
\arguments{
\item{dim_names}{A character vector of dimension names.}
}
\value{
A character vector of unique dimension names.
}
\description{
Process Dimension Names to Handle Duplicates
}
