% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pattern}
\alias{process_pattern}
\title{Process a pattern within a data object}
\usage{
process_pattern(pattern, data_obj, exp_name, pattern_mix = FALSE)
}
\arguments{
\item{pattern}{A character string representing the pattern to match.}

\item{data_obj}{A data object containing dimension information and data.}

\item{exp_name}{A character string representing the experiment name.}

\item{pattern_mix}{Logical; if TRUE, allows matching of mixed patterns.}
}
\value{
A data frame containing processed data for the matching pattern, or NULL if no matches are found.
}
\description{
Extracts and processes matching variables from a data object based on a given pattern.
}
\keyword{internal}
