% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_sl4col}
\alias{rename_sl4col}
\title{Rename Columns (Internal)}
\usage{
rename_sl4col(df, rename_cols)
}
\arguments{
\item{df}{A data frame.}

\item{rename_cols}{A named vector where names represent old column names and values represent new names.}
}
\value{
A data frame with renamed columns.
}
\description{
A helper function that renames columns inside the function.
}
\keyword{internal}
