% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sl4x.R
\name{rename_sl4col_all}
\alias{rename_sl4col_all}
\title{Rename Data Frame Columns Based on Mapping}
\usage{
rename_sl4col_all(df, mapping_df)
}
\arguments{
\item{df}{A data frame whose columns need to be renamed.}

\item{mapping_df}{A two-column data frame where the first column contains old names,
and the second column contains new names.}
}
\value{
A data frame with renamed columns.
}
\description{
Renames the columns of a data frame according to a specified mapping,
ensuring that duplicate names are avoided by appending numeric suffixes if necessary.
}
