% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_pattern}
\alias{standardize_pattern}
\title{Standardize dimension pattern}
\usage{
standardize_pattern(pattern)
}
\arguments{
\item{pattern}{Dimension pattern string}
}
\value{
Standardized dimension pattern
}
\description{
Standardize dimension pattern
}
