% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sl4x.R
\name{get_dims_elements}
\alias{get_dims_elements}
\title{Extract Unique Dimension Elements}
\usage{
get_dims_elements(sl4_obj)
}
\arguments{
\item{sl4_obj}{A structured SL4 object containing dimension information.}
}
\value{
A data frame containing unique dimension elements.
}
\description{
Extracts and lists unique dimension elements present in an SL4 object.
}
